Build: restructure inner folders so that llm, stt, and tts are of equivalent(ish) design
