+++ stt/src/microphone_recognition.py
