+++ stt/src/simple_vosk.py
